+++ tests/utils.js
@@ -44,4 +44,6 @@ export function reset() {
 }
 
 // Mock the hash function.
-jss.generateClassName = (str, rule) => (rule.name ? `${rule.name}-id` : 'id')
+jss.setup({
+  generateClassName: (str, rule) => (rule.name ? `${rule.name}-id` : 'id')
+})